refactor handler setup and main.go
