+++ server/data/db.go
@@ -11,7 +11,6 @@ import (
 var db *pgxpool.Pool
 
 func Connect() {
-
 	name := os.Getenv("POSTGRES_DB")
 	password := os.Getenv("POSTGRES_PASSWORD")
 	host := os.Getenv("POSTGRES_HOST")
